refactor: extract progressionAggregation module from calculator (Phase 4a.5)
Move aggregateProgressionGains, aggregatePathBenefits, and
checkFlavorFeatureAutoGrant into calculatorModules/progressionAggregation.ts.
All snapshot tests pass with identical output.
