+++ src/lib/services/enhancedCharacterCalculator.ts
@@ -24,8 +24,7 @@ import {
 	type IncreaseMasteryCapEffect
 } from '../rulesdata/schemas/character.schema';
 import {
-	resolveClassProgression,
-	resolveSubclassFeatures
+	resolveClassProgression
 } from '../rulesdata/classes-data/classProgressionResolver';
 import { classesData } from '../rulesdata/loaders/class.loader';
 // findTalentById moved to calculatorModules/effectCollection.ts
@@ -53,7 +52,7 @@ import { attributesData } from '../rulesdata/attributes';
 import { skillsData } from '../rulesdata/skills';
 import { tradesData } from '../rulesdata/trades';
 import type { ClassDefinition } from '../rulesdata/schemas/character.schema';
-import { CHARACTER_PATHS } from '../rulesdata/progression/paths/paths.data';
+// CHARACTER_PATHS moved to calculatorModules/progressionAggregation.ts
 import {
 	collectGrantedAbilities,
 	collectMovements,
@@ -72,189 +71,12 @@ import {
 	aggregateAttributedEffects,
 	resolveEffectChoices
 } from './calculatorModules/effectCollection';
+import {
+	aggregateProgressionGains,
+	checkFlavorFeatureAutoGrant
+} from './calculatorModules/progressionAggregation';
 
-/**
- * Cross-path grants returned by aggregatePathBenefits (DC20 v0.10 p.161)
- */
-interface CrossPathGrants {
-	/** Spellcaster taking Martial Path gets Spellcaster Stamina Regen */
-	grantsSpellcasterStaminaRegen: boolean;
-	/** Martial taking Spellcaster Path must choose a Spell List */
-	requiresSpellListChoice: boolean;
-	/** Combat Training grants from path progression */
-	combatTrainingGrants: string[];
-}
-
-/**
- * Helper to aggregate path point benefits based on allocations
- * @param pathPointAllocations - How many points allocated to each path
- * @param classCategory - The class's category (martial/spellcaster/hybrid) for cross-path rules
- * @returns Aggregated benefits from all allocated path levels, including cross-path grants
- */
-function aggregatePathBenefits(
-	pathPointAllocations?: {
-		martial?: number;
-		spellcasting?: number;
-	},
-	classCategory?: 'martial' | 'spellcaster' | 'hybrid'
-): {
-	totalSP: number;
-	totalMP: number;
-	totalManeuversKnown: number;
-	totalCantripsKnown: number;
-	totalSpellsKnown: number;
-	crossPathGrants: CrossPathGrants;
-} {
-	let totalSP = 0;
-	let totalMP = 0;
-	let totalManeuversKnown = 0;
-	let totalCantripsKnown = 0;
-	let totalSpellsKnown = 0;
-
-	// Initialize cross-path grants
-	const crossPathGrants: CrossPathGrants = {
-		grantsSpellcasterStaminaRegen: false,
-		requiresSpellListChoice: false,
-		combatTrainingGrants: []
-	};
-
-	if (!pathPointAllocations) {
-		return {
-			totalSP,
-			totalMP,
-			totalManeuversKnown,
-			totalCantripsKnown,
-			totalSpellsKnown,
-			crossPathGrants
-		};
-	}
-
-	// Martial Path bonuses
-	if (pathPointAllocations.martial && pathPointAllocations.martial > 0) {
-		const martialPath = CHARACTER_PATHS.find((p) => p.id === 'martial_path');
-		if (martialPath) {
-			for (let level = 1; level <= pathPointAllocations.martial; level++) {
-				const levelData = martialPath.progression.find((p) => p.pathLevel === level);
-				if (levelData?.benefits) {
-					totalSP += levelData.benefits.staminaPoints || 0;
-					totalManeuversKnown += levelData.benefits.maneuversLearned || 0;
-				}
-			}
-
-			// DC20 v0.10 p.161: Grant Combat Training with Weapons from Martial Path
-			crossPathGrants.combatTrainingGrants.push('Weapons');
-
-			// DC20 v0.10 p.161: Spellcaster classes (lacking Stamina Regen) gain
-			// "Spellcaster Stamina Regen" when they first invest in Martial Path
-			if (classCategory === 'spellcaster') {
-				crossPathGrants.grantsSpellcasterStaminaRegen = true;
-				console.log('✨ Cross-path grant: Spellcaster Stamina Regen', {
-					classCategory,
-					martialPoints: pathPointAllocations.martial
-				});
-			}
-		}
-	}
-
-	// Spellcaster Path bonuses
-	if (pathPointAllocations.spellcasting && pathPointAllocations.spellcasting > 0) {
-		const spellcasterPath = CHARACTER_PATHS.find((p) => p.id === 'spellcaster_path');
-		if (spellcasterPath) {
-			for (let level = 1; level <= pathPointAllocations.spellcasting; level++) {
-				const levelData = spellcasterPath.progression.find((p) => p.pathLevel === level);
-				if (levelData?.benefits) {
-					totalMP += levelData.benefits.manaPoints || 0;
-					totalCantripsKnown += levelData.benefits.cantripsLearned || 0;
-					totalSpellsKnown += levelData.benefits.spellsLearned || 0;
-				}
-			}
-
-			// DC20 v0.10 p.161: Grant Combat Training with Spell Focuses from Spellcaster Path
-			crossPathGrants.combatTrainingGrants.push('Spell_Focuses');
-
-			// DC20 v0.10 p.161: Martial classes (lacking a Spell List) gain
-			// a Spell List of their choice when they first invest in Spellcaster Path
-			if (classCategory === 'martial') {
-				crossPathGrants.requiresSpellListChoice = true;
-				console.log('✨ Cross-path grant: Requires Spell List choice', {
-					classCategory,
-					spellcastingPoints: pathPointAllocations.spellcasting
-				});
-			}
-		}
-	}
-
-	return {
-		totalSP,
-		totalMP,
-		totalManeuversKnown,
-		totalCantripsKnown,
-		totalSpellsKnown,
-		crossPathGrants
-	};
-}
-
-/**
- * Check if a class's Flavor Feature should be auto-granted (DC20 v0.10 p.161)
- * Rule: "Once you gain 2 Class Features from the same Class, you automatically gain that Class's Flavor Feature."
- *
- * @param classId - The class to check
- * @param unlockedFeatureIds - Feature IDs unlocked from main class progression
- * @param multiclassFeatures - Array of {classId, featureId} from multiclass talents
- * @returns The flavor feature to auto-grant, or null if criteria not met
- */
-function checkFlavorFeatureAutoGrant(
-	classId: string,
-	unlockedFeatureIds: string[],
-	multiclassFeatures: Array<{ classId: string; featureId: string }>
-): { featureId: string; featureName: string; classId: string } | null {
-	const classDefinition = findClassByName(classId);
-	if (!classDefinition) return null;
-
-	// Count non-flavor features from this class
-	const classFeatures = classDefinition.coreFeatures.filter((f) => !f.isFlavor);
-
-	// Count features unlocked from main class progression
-	let featureCount = 0;
-	for (const feature of classFeatures) {
-		const featureId = feature.id || feature.featureName;
-		if (unlockedFeatureIds.includes(featureId)) {
-			featureCount++;
-		}
-	}
-
-	// Count features from multiclass (if targeting this class)
-	for (const mc of multiclassFeatures) {
-		if (mc.classId === classId) {
-			// Verify it's not a flavor feature
-			const feature = classFeatures.find((f) => (f.id || f.featureName) === mc.featureId);
-			if (feature && !feature.isFlavor) {
-				featureCount++;
-			}
-		}
-	}
-
-	console.log('🎭 Flavor Feature check:', {
-		classId,
-		featureCount,
-		threshold: 2,
-		qualifies: featureCount >= 2
-	});
-
-	// If 2+ features, find and return the flavor feature
-	if (featureCount >= 2) {
-		const flavorFeature = classDefinition.coreFeatures.find((f) => f.isFlavor);
-		if (flavorFeature) {
-			return {
-				featureId: flavorFeature.id || flavorFeature.featureName,
-				featureName: flavorFeature.featureName,
-				classId
-			};
-		}
-	}
-
-	return null;
-}
+// CrossPathGrants, aggregatePathBenefits, checkFlavorFeatureAutoGrant moved to calculatorModules/progressionAggregation.ts
 
 /**
  * Convert character context data to enhanced build data
@@ -476,189 +298,7 @@ function getOptionsForEffect(effect: AttributedEffect): ChoiceOption[] {
 	}));
 }
 
-/**
- * Aggregate all progression gains from level 1 to target level
- */
-function aggregateProgressionGains(
-	classProgressionData: any,
-	targetLevel: number,
-	pathPointAllocations?: { martial?: number; spellcasting?: number },
-	selectedSubclass?: string,
-	classCategory?: 'martial' | 'spellcaster' | 'hybrid'
-): {
-	totalHP: number;
-	totalSP: number;
-	totalMP: number;
-	totalSkillPoints: number;
-	totalTradePoints: number;
-	totalAttributePoints: number;
-	totalManeuversKnown: number;
-	totalCantripsKnown: number;
-	totalSpellsKnown: number;
-	totalTalents: number;
-	totalPathPoints: number;
-	totalAncestryPoints: number;
-	unlockedFeatureIds: string[];
-	pendingSubclassChoices: number;
-	crossPathGrants: CrossPathGrants;
-} {
-	let totalHP = 0;
-	let totalSP = 0;
-	let totalMP = 0;
-	let totalSkillPoints = 0;
-	let totalTradePoints = 0;
-	let totalAttributePoints = 0;
-	let totalManeuversKnown = 0;
-	let totalCantripsKnown = 0;
-	let totalSpellsKnown = 0;
-	let totalTalents = 0;
-	let totalPathPoints = 0;
-	let totalAncestryPoints = 0;
-	const unlockedFeatureIds: string[] = [];
-	let pendingSubclassChoices = 0;
-
-	if (!classProgressionData?.levelProgression) {
-		return {
-			totalHP,
-			totalSP,
-			totalMP,
-			totalSkillPoints,
-			totalTradePoints,
-			totalAttributePoints,
-			totalManeuversKnown,
-			totalSpellsKnown,
-			totalTalents,
-			totalPathPoints,
-			totalAncestryPoints,
-			unlockedFeatureIds,
-			pendingSubclassChoices
-		};
-	}
-
-	console.log('🔍 AGGREGATION START:', {
-		classId: classProgressionData?.id,
-		targetLevel,
-		hasProgression: !!classProgressionData?.levelProgression,
-		progressionLength: classProgressionData?.levelProgression?.length
-	});
-
-	for (let level = 1; level <= targetLevel; level++) {
-		const levelData = classProgressionData.levelProgression.find((lp: any) => lp.level === level);
-		if (!levelData) {
-			console.warn(`⚠️ Level ${level} not found in progression!`);
-			continue;
-		}
-
-		console.log(`📊 Level ${level} data:`, {
-			hasGains: !!levelData.gains,
-			gains: JSON.stringify(levelData.gains),
-			talents: levelData.gains?.talents,
-			pathPoints: levelData.gains?.pathPoints
-		});
-
-		// Aggregate numeric stats from progression files (gained* fields)
-		totalHP += levelData.gainedHealth || levelData.healthPoints || 0;
-		totalSP += levelData.gainedStaminaPoints || levelData.staminaPoints || 0;
-		totalMP += levelData.gainedManaPoints || levelData.manaPoints || 0;
-		totalSkillPoints += levelData.gainedSkillPoints || levelData.skillPoints || 0;
-		totalTradePoints += levelData.gainedTradePoints || levelData.tradePoints || 0;
-		totalAttributePoints += levelData.gainedAttributePoints || levelData.attributePoints || 0;
-		totalManeuversKnown += levelData.gainedManeuversKnown || levelData.maneuversKnown || 0;
-		totalSpellsKnown +=
-			(levelData.gainedSpellsKnown || levelData.spellsKnown || 0) +
-			(levelData.gainedCantripsKnown || levelData.cantripsKnown || 0);
-
-		// Aggregate new structured gains (if present)
-		if (levelData.gains) {
-			totalTalents += levelData.gains.talents || 0;
-			// Support both pathPoints (number) and pathProgression (boolean).
-			totalPathPoints += levelData.gains.pathPoints || (levelData.gains.pathProgression ? 1 : 0);
-			totalAncestryPoints += levelData.gains.ancestryPoints || 0;
-
-			if (levelData.gains.classFeatures) {
-				unlockedFeatureIds.push(...levelData.gains.classFeatures);
-			}
-
-			if (levelData.gains.subclassFeatureChoice) {
-				pendingSubclassChoices++;
-			}
-		}
-	}
-
-	console.log('✅ AGGREGATION COMPLETE (before path bonuses):', {
-		totalTalents,
-		totalPathPoints,
-		totalSkillPoints,
-		totalTradePoints,
-		unlockedFeatureIds: unlockedFeatureIds.length,
-		totalSP,
-		totalMP,
-		totalManeuversKnown
-	});
-
-	// Add path bonuses to progression totals (with cross-path grants)
-	const pathBonuses = aggregatePathBenefits(pathPointAllocations, classCategory);
-	totalSP += pathBonuses.totalSP;
-	totalMP += pathBonuses.totalMP;
-	totalManeuversKnown += pathBonuses.totalManeuversKnown;
-	totalCantripsKnown += pathBonuses.totalCantripsKnown;
-	totalSpellsKnown += pathBonuses.totalSpellsKnown;
-
-	console.log('✅ PATH BONUSES APPLIED:', {
-		pathAllocations: pathPointAllocations,
-		classCategory,
-		bonuses: pathBonuses,
-		crossPathGrants: pathBonuses.crossPathGrants,
-		finalSP: totalSP,
-		finalMP: totalMP,
-		finalManeuvers: totalManeuversKnown
-	});
-
-	// Apply subclass features (M3.10c)
-	if (selectedSubclass && classProgressionData?.id) {
-		try {
-			const subclassFeatures = resolveSubclassFeatures(
-				classProgressionData.id,
-				selectedSubclass,
-				targetLevel
-			);
-
-			console.log('🎭 SUBCLASS FEATURES APPLIED:', {
-				subclass: selectedSubclass,
-				featuresCount: subclassFeatures.length,
-				featureIds: subclassFeatures.map((f) => f.id || f.featureName)
-			});
-
-			// Add subclass feature IDs to unlocked features
-			subclassFeatures.forEach((feature) => {
-				const featureId = feature.id || feature.featureName;
-				if (featureId) {
-					unlockedFeatureIds.push(featureId);
-				}
-			});
-		} catch (error) {
-			console.error('Failed to resolve subclass features:', error);
-		}
-	}
-
-	return {
-		totalHP,
-		totalSP,
-		totalMP,
-		totalSkillPoints,
-		totalTradePoints,
-		totalAttributePoints,
-		totalManeuversKnown,
-		totalCantripsKnown,
-		totalSpellsKnown,
-		totalTalents,
-		totalPathPoints,
-		totalAncestryPoints,
-		unlockedFeatureIds,
-		pendingSubclassChoices,
-		crossPathGrants: pathBonuses.crossPathGrants
-	};
-}
+// aggregateProgressionGains moved to calculatorModules/progressionAggregation.ts
 
 /**
  * Main calculation function with detailed breakdowns
